requirements.txt
# Core dependencies for tf6d
# Note: PyTorch, torchvision, torchaudio, and faiss-gpu should be installed
# via conda first. See README.md for full setup instructions.

tqdm==4.65.0
timm==0.9.16
matplotlib==3.8.3
scikit-learn==1.4.1.post1
opencv-python==4.10.0.82
pydensecrf @ git+https://github.com/lucasb-eyer/pydensecrf.git@dd070546eda51e21ab772ee6f14807c7f5b1548b
transforms3d==0.4.1
pillow==9.4.0
plyfile==1.0.3
trimesh==4.1.4
imageio==2.34.0
pypng==0.20220715.0
vispy==0.12.2
pyopengl==3.1.1a1
pyglet==2.0.10
numba==0.59.0
hydra-core>=1.1
loguru>=0.3
numpy>=1.17
omegaconf>=2.1
albumentations==1.3.1
open_clip_torch==2.23.0
einops==0.7.0
segment-anything @ git+https://github.com/facebookresearch/segment-anything.git
transformers==4.29.2
diffusers==0.16.1
accelerate==0.20.1
pytorch-lightning==1.4.2
torchmetrics==0.6.0
kornia==0.6
mat73==0.62
pandas
seaborn
protobuf==3.20.3
poselib
